Initial gui implementation
Up until this point we have been rendering our video in a raw opengl
context with a glfw window. The plan has always been to move to a real
gui framework, but keep the existing rendering code in place. Well here
we are.

The GUI framework choice is a little odd. Pulling in rust as a
dependency does not make nearly as much sense as using something that
could be compiled with zig directly (any c/c++ library). Unfortunately,
we are going to have some pretty strict requirements for our editor.
When we eventually try to lay out a script for the video, we are going
to want to

* Jump to position in video based off which word was clicked
* Add some indicator to the text that represents where the current video
  position is
* Add indicators for segments of video that are planned to be
  imported/exported

Some of these are pretty hard to get. I skimmed through Qt, nuklear, and
Dear ImGui docs, and I couldn't see how to easily pre-layout text and
then paint my own stuff on top. egui (which I've used before) has this
ability through it's Galley type. So here we are, pulling in a rust
dependency for our zig project

This seems heavy, and it's surprising that egui feels like the best
choice for this setup. A goal of this implementation is to try to keep
the GUI code as light as possible, with the possibility of switching GUI
frameworks (and even languages) in the future. Because of this we want
to keep as much code in zig land as possible

* Implement a GuiGl bindings type. This allows the GUI to re-export it's
  opengl bindings in a way that looks familiar to what the zig side was
  already using
* Add a single gui_run() function that starts the GUI mainloop
* Add C exports on the zig side for a FrameRenderer. This is the old Gui
  object, but stripped down to only handle frame rendering
* With the current setup, we do not have any ability to interact with
  the GUI from the main loop, except through the FrameRenderer object.
  This will come in the future, but for now this means
  * egui is not aware when new frames come in, and therefore doesn't
    know when to redraw
  * We have lost the ability to pause
  * Our linter script cannot easily exit
  * This could be fixed, however in the interests of getting something
    committed without stream missing anything important, we will make do
    with what we have now
* Valgrind usage with egui goes haywire. It's slow, reports lots of
  issues, and we have no ability to exit the lint script early. To
  address all of these issues we've implemented a mock ui that is used
  in our linter.
  * In this setup we track GuiGl calls to create/destroy
    programs/shaders/textures through a simple malloc. This allows
    valgrind to let us know if we forget to free something that we were
    supposed to