Fix frames not updating
As an optimization, egui only redraws when it thinks the state has
changed in some way. This means that if we swap out a video frame, we
will not see it until we interact with the window in some way.

For our case, egui provides a request_repaint() function on it's
context. We need to request a repaint every time we push a new frame.
This means that in zig land, we need a handle to the egui context.

Adjust our GUI api to look more like a typical C object. Return a gui
handle on init, and use that state when we run the gui. This handle
is given an egui context when the UI thread actually starts running

Adjust our GuiGlImpl in our mock UI to be shared as a GuiImpl

Since there is an exported Gui ffi concept, rename the old gui
EframeImpl to free up the name