Re-implement pause with new UI framework
In order to implement pause, we need a way for the GUI to inform the
main loop about events, and vice versa.

Add "actions" that the GUI can emit. These are implemented as a channel
on the rust side that is polled from zig through the FFI

Add an AppState object that returns AppStateSnapshots from the zig side.
When egui redraws, we ask zig for a snapshot of the current app state.
This currently only keeps track of if we're paused, but will be extended
to provide more app context in the future.

When paused, the audio thread is not happy about the lack of audio
frames. Remove the error log for now as this is expected behavior

Bonus fix: fix the app name