Close GUI on main loop crash
Previously, if the main loop exited, the gui thread would just sit there
until it was closed. The main loop should have some way of requesting
that the GUI is closed.

Add an API to close the UI from the app. Note that since the main loop
can fail early, we need to ensure that the gui side is ready to receive
the close request before we start the main thread, otherwise the gui
will miss the shutdown notification