Fix confusing memory deinit
The video decoder hands out frame references that require the decoder to
have a stable memory location. When we deinit the frame renderer shared
data, it may free a cached frame.

Because of this the shared renderer data has to be destructed after the
decoder, which resulted in the deinit() being spaced far from init() in
the previous impl

Move the decoder to the beginning of main, ensuring that frame
references should always point to somewhere valid. Then move the shared
frame data deinit() to be near its init(), as god intended.