Split out app object
It's finally time. The main loop was starting to get kinda hairy, and
splitting out functions was difficult because there was so much state to
pass around. E.g. a seekToPts() function requires

* Decoder
* Gui
* Frame renderer
* Player state
* Last pts
* Stream ID
* Current time
* Next frame pts

There is just too much intertwined state to pretty up without splitting
out an object

Take what used to be the main loop, and throw it all into a struct, then
update our main loop to be more split into manageable chunks.

While we're here, move the app class into it's own file so that there's
less to grok in main