Lint script performance improvement
With the initial audio renderer implementation, the performance under
valgrind was unbearably bad. It seems that we're essentially hot looping
on re-populaiting the vertex buffer, and in the process running a lot of
unnecessary debug assertions in array list. Replacing the array list
with a raw array works much better

While we're here, turn on the optimizer for running under valgrind, why
not

And use the C allocator in release mode, it's faster than the general
purpose allocator