Fix incorrect audio extraction from ffmpeg
We assumed that the linesize indicated the correct size of data. This
was a bad assumption. In some cases this was producing choppy audio if
we tried to use the data directly

Luckily when consuming the audio in miniaudio, we happened to be
checking the number of samples as well, but in new code this caused
problems