Add seeking based off auto generated subtitles
This is the major feature of this whole project. Generate subtitles for
the input video, show them on the side of the editor, and jump to where
that word was said on click. There are plenty of improvements to be made
here, but this is a good milestone for a V0.00

* Add whisper.cpp as a dependency
* On init, use whisper.cpp to generate a word timestamp mapping
  * Resample audio down to 16000Hz as required by the whisper model,
    only take 1 channel of first audio stream
  * Use quantized, tiny, english model for now. We can always
    upgrade/give options later
* Feed the text to the UI for display
* When text is clicked, look up the associated timestamp from the word
  timestamp map, and use to request a seek from the application
* Subtitle generation is quite expensive to the point of being
  unbearable in valgrind. Add runtime flag to disable