Fix seeking
It seems that AVCodecs can have some buffered state that causes issues
when seeking. The docs for avcodec_flush_buffers() say we should call it
after seeking, so we can do that. This fixes undesirable behavior where
sometimes when seeking backwards, the rest of the system would think
that we hadn't done a seek, resulting in maxing CPU decoding frames from
the desired position to where we came from

After the above fix, there is still an issue where seeking can take a
significant amount of time. I was seeing up to 500ms in some cases. This
causes us to fall behind, and when we try to catch up we hammer the
audio thread too hard which is nasty to listen to. Update the current
time after we seek so we have an accurate account of when we decoded the
frame we seeked to