Disable debug symbols for rust build
Unfortunate but maybe useful evil. With debug symbols available, we get
a 200MB output binary. This feels like it should be fine, but there are
a couple levels of caching/copying that are done while generating this.
It turns out that if we are in a rapid edit/compile loop (which is very
possible), we can end up thrashing our disk to the point where it takes
over a minute to do what needs to be done.

With debug info off, we get that down to 65MBish

With the fake UI it's down to 20MBish, down to 5ish if we strip

Disable debug symbols for now, and we can turn them on temporarily if we
need them in the future