Implement drag seeking on script
Seeking behavior on the timeline should be consistent with the seeking
behavior on the script. Drag seeking is nicer than click. Move the
pause/unpause handling out to it's own struct and pass everywhere that
needs it. That makes it simple to just enable drag seeking on the script
widget