Save/resume script generation
Script generation can take a long time. If we've spent an hour running
whisper, we should save those results somewhere. For now just stash it
in our save file

As part of this the ownership of saving/loading has to be moved out of
ClipManager. Since this is a more widespread concept, hide the fact that
the script is json behind save module.

std.json has a way of parsing a json blob directly into a struct,
however it does not give very good error messages (or any error
messages) on failure. Re-implement parsing around std.json.Value so that
we can flag exactly what is wrong with the save